fix refund amount
