+++ src/VirtoCommerce.CyberSourcePayment.Core/Models/CyberSourceCapturePaymentRequest.cs
@@ -2,7 +2,11 @@ using VirtoCommerce.OrdersModule.Core.Model;
 
 namespace VirtoCommerce.CyberSourcePayment.Core.Models;
 
-public class CyberSourceRefundPaymentRequest : CyberSourceRequest
+public class CyberSourceCapturePaymentRequest : CyberSourceRequest
 {
-    public PaymentIn Payment { get; set; }
+    public CustomerOrder Order { get; set; }
+    public decimal? Amount { get; set; }
+    public int PaymentNumber { get; set; }
+    public bool IsFinal { get; set; }
+    public string Notes { get; set; }
 }
