+++ src/VirtoCommerce.CyberSourcePayment.Core/Models/CyberSourceProcessPaymentRequest.cs
@@ -5,6 +5,5 @@ namespace VirtoCommerce.CyberSourcePayment.Core.Models;
 public class CyberSourceProcessPaymentRequest : CyberSourceRequest
 {
     public string Token { get; set; }
-    public PaymentIn Payment { get; set; }
     public CustomerOrder Order { get; set; }
 }
