+++ src/VirtoCommerce.CyberSourcePayment.Core/Models/CyberSourceVoidPaymentRequest.cs
@@ -4,6 +4,5 @@ namespace VirtoCommerce.CyberSourcePayment.Core.Models;
 
 public class CyberSourceVoidPaymentRequest : CyberSourceRequest
 {
-    public PaymentIn Payment { get; set; }
     public CustomerOrder Order { get; set; }
 }
